tests/submission/Infraware/Forms/contents/Forms/output_defaultvalue.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>output_defaultValue</h3>
	</p>
	
	<hr>

	<div id="log"></div>
	
	<form onsubmit="return false" name='input_form'>
	 value: <output name=o id='output_id'></output>
	</form>

	<script>
		
		var output = document.getElementById("output_id");
		output.defaultValue = 'val';

		if (typeof(output.defaultValue) == "string") {
			test(function() {
				assert_equals(output.defaultValue, "val", "defaultValue attribute is not exist.");
			});
		} else {
			test(function() {
				assert_unreached("defaultValue attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
